update
- load format versions from auto-completion engine
- better support for using the auto-completion engine as a database
- add icon for setting project target version
- restructured VersionedTemplate code
